+++ packages/salesforcedx-vscode-core/src/commands/index.ts
@@ -76,7 +76,6 @@ export {
   ForceListMetadataExecutor
 } from './forceListMetadata';
 export {
-  forceSourceRetrieve,
+  forceSourceRetrieveCmp,
   ForceSourceRetrieveExecutor
-} from './forceSourceRetrieve';
-export { forceSourceRetrieveCmp } from './forceSourceRetrieveCmp';
+} from './forceSourceRetrieveCmp';
